Check for valid job ids
The sacct command seems to freeze when passed an invalid job id like
'sacct -j X', so I try and validate the input. We need to allow
underscores and dots in case someone wants to query part of a job array
or a single sub step of a job.

I am not 100% sure this covers all cases, but probably better than no
checks.