Only hightlight complete matches
Before all the search patterns were used directly, which meant that 'a'
would match all users/nodes with an a in their name.
Now each pattern is surrounded by ^ and $ forcing it to make a complete
match. It is still a regex, so 's03n.*' can be used to match all nodes
starting with s03, etc.