Show 0 memory etc when reported by sacct
Jobinfo was showing '--' even when `sacct` reports a memory usage of 0
and not just when we cannot get the usage at all.

Also makes the types a bit less stringly for some of the fields -
couldn't get it into a separate commit in the time I was willing to put
in.