Show ExitCode field
FIXME: For jobs that haven't completed yet, Slurm reports this value
as '0:0', which I find a bit incorrect (there can't be an _exit code_
if there hasn't been an _exit_ yet, no?).  At the moment the script
reports Slurm's givings verbatim, but perhaps a '--' would've been
better for such cases.